+++ forgejo/scripts/import.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
-for dir in /Users/tomwolfe/Development/SIMPLIFi/*/; do
+forgeBase="ssh://git@macmini.local:2222"
+
+for dir in "$1"/*/*/; do
+  org=$(dirname "$dir")
+  org=$(basename "$org")
   name=$(basename "$dir")
-  echo $name
-  git -C "$dir" remote set-url origin "ssh://git@macmini.local:2222/SIMPLIFi/$name.git"
-  git -C "$dir" push
+  echo "$forgeBase/$org/$name.git"
+  git -C "$dir" remote remove origin
+  git -C "$dir" remote add origin "$forgeBase/$org/$name.git"
+  git -C "$dir" push origin
 done